nightly-unstable/xhtml1/events-006.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
<meta charset="utf-8" />
<title>CSS Transitions Test: transitionend event for pseudo elements</title>
<meta content="Test checks that TransitionEnd events are fired for pseudo-elements" name="assert" />
<link href="http://www.w3.org/TR/css3-transitions/#transition-events" rel="help" title="5. Transition Events" />
<link href="http://www.w3.org/TR/CSS21/generate.html#before-after-content" rel="help" title="CSS21 - 12.1 The :before and :after pseudo-elements" />
<link href="http://www.w3.org/TR/css3-content/" rel="help" title="CSS3 Generated and Replaced Content Module" />
<link href="http://rodneyrehm.de/en/" rel="author" title="Rodney Rehm" />

<script src="/resources/testharness.js" type="text/javascript"></script>
<script src="/resources/testharnessreport.js" type="text/javascript"></script>
<script src="./support/helper.js" type="text/javascript"></script>

<style>
.before::before,
.after:after {
  content: '';
  transition: padding-left .01s;
  padding-left: 1px;
}
.before.active::before,
.after.active:after {
  padding-left: 10px;
}
</style>

</head>
<body>

<div id="log"></div>

<script>
promise_test(t =&gt; {
  const div = addDiv(t, { 'class': 'before' });
  getComputedStyle(div).paddingLeft;
  div.classList.add('active');

  const watcher = new EventWatcher(t, div, [ 'transitionend' ]);
  return watcher.wait_for('transitionend').then(evt =&gt; {
    assert_end_events_equal(evt, 'padding-left', 0.01, '::before');
  });
}, 'transition padding-left on ::before');

promise_test(t =&gt; {
  const div = addDiv(t, { 'class': 'after' });
  getComputedStyle(div).paddingLeft;
  div.classList.add('active');

  const watcher = new EventWatcher(t, div, [ 'transitionend' ]);
  return watcher.wait_for('transitionend').then(evt =&gt; {
    assert_end_events_equal(evt, 'padding-left', 0.01, '::after');
  });
}, 'transition padding-left on ::after');
</script>


</body></html>
